requirements.txt
hopsworks==4.6.0
kagglehub==0.3.8
pandas
numpy
scikit-learn
xgboost
python-dotenv
streamlit
plotly
